r/aws_backup_vault: Reduce visibility.
