+++ internal/service/backup/vault.go
@@ -20,6 +20,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/hashicorp/terraform-provider-aws/internal/conns"
+	"github.com/hashicorp/terraform-provider-aws/internal/enum"
 	"github.com/hashicorp/terraform-provider-aws/internal/errs"
 	"github.com/hashicorp/terraform-provider-aws/internal/errs/sdkdiag"
 	tftags "github.com/hashicorp/terraform-provider-aws/internal/tags"
@@ -30,7 +31,7 @@ import (
 
 // @SDKResource("aws_backup_vault", name="Vault")
 // @Tags(identifierAttribute="arn")
-func ResourceVault() *schema.Resource {
+func resourceVault() *schema.Resource {
 	return &schema.Resource{
 		CreateWithoutTimeout: resourceVaultCreate,
 		ReadWithoutTimeout:   resourceVaultRead,
@@ -151,7 +152,6 @@ func resourceVaultDelete(ctx context.Context, d *schema.ResourceData, meta inter
 		var errs []error
 
 		pages := backup.NewListRecoveryPointsByBackupVaultPaginator(conn, input)
-
 		for pages.HasMorePages() {
 			page, err := pages.NextPage(ctx)
 
@@ -173,12 +173,12 @@ func resourceVaultDelete(ctx context.Context, d *schema.ResourceData, meta inter
 				})
 
 				if err != nil {
-					errs = append(errs, fmt.Errorf("deleting recovery point (%s): %w", recoveryPointARN, err))
+					errs = append(errs, fmt.Errorf("deleting Backup Vault recovery point (%s): %w", recoveryPointARN, err))
 					continue
 				}
 
 				if _, err := waitRecoveryPointDeleted(ctx, conn, d.Id(), recoveryPointARN, d.Timeout(schema.TimeoutDelete)); err != nil {
-					errs = append(errs, fmt.Errorf("waiting for recovery point (%s) delete: %w", recoveryPointARN, err))
+					errs = append(errs, fmt.Errorf("waiting for Backup Vault recovery point (%s) delete: %w", recoveryPointARN, err))
 					continue
 				}
 			}
@@ -242,4 +242,69 @@ func findVault(ctx context.Context, conn *backup.Client, input *backup.DescribeB
 	}
 
 	return output, nil
+}
+
+func findRecoveryPointByTwoPartKey(ctx context.Context, conn *backup.Client, backupVaultName, recoveryPointARN string) (*backup.DescribeRecoveryPointOutput, error) {
+	input := &backup.DescribeRecoveryPointInput{
+		BackupVaultName:  aws.String(backupVaultName),
+		RecoveryPointArn: aws.String(recoveryPointARN),
+	}
+
+	return findRecoveryPoint(ctx, conn, input)
+}
+
+func findRecoveryPoint(ctx context.Context, conn *backup.Client, input *backup.DescribeRecoveryPointInput) (*backup.DescribeRecoveryPointOutput, error) {
+	output, err := conn.DescribeRecoveryPoint(ctx, input)
+
+	if errs.IsA[*awstypes.ResourceNotFoundException](err) {
+		return nil, &retry.NotFoundError{
+			LastError:   err,
+			LastRequest: input,
+		}
+	}
+
+	if err != nil {
+		return nil, err
+	}
+
+	if output == nil {
+		return nil, tfresource.NewEmptyResultError(input)
+	}
+
+	return output, nil
+}
+
+func statusRecoveryPoint(ctx context.Context, conn *backup.Client, backupVaultName, recoveryPointARN string) retry.StateRefreshFunc {
+	return func() (interface{}, string, error) {
+		output, err := findRecoveryPointByTwoPartKey(ctx, conn, backupVaultName, recoveryPointARN)
+
+		if tfresource.NotFound(err) {
+			return nil, "", nil
+		}
+
+		if err != nil {
+			return nil, "", err
+		}
+
+		return output, string(output.Status), nil
+	}
+}
+
+func waitRecoveryPointDeleted(ctx context.Context, conn *backup.Client, backupVaultName, recoveryPointARN string, timeout time.Duration) (*backup.DescribeRecoveryPointOutput, error) {
+	stateConf := &retry.StateChangeConf{
+		Pending: enum.Slice(awstypes.RecoveryPointStatusDeleting),
+		Target:  []string{},
+		Refresh: statusRecoveryPoint(ctx, conn, backupVaultName, recoveryPointARN),
+		Timeout: timeout,
+	}
+
+	outputRaw, err := stateConf.WaitForStateContext(ctx)
+
+	if output, ok := outputRaw.(*backup.DescribeRecoveryPointOutput); ok {
+		tfresource.SetLastError(err, errors.New(aws.ToString(output.StatusMessage)))
+
+		return output, err
+	}
+
+	return nil, err
 }
